+++ resources/js/app.js
@@ -367,12 +367,12 @@ $('.item-quantity').on('change', function(){
 
 $( ".item-option-select" ).on('change', function(){
 
-  var basePrice = parseInt($('#base_price').text());
+  var basePrice = parseFloat($('#base_price').text());
   var optionTotal = 0.00;
 
   $( ".item-option-select" ).each(function( index ) {
     if($(this).find(":selected").data('price-mod')){
-      optionTotal += parseInt($(this).find(":selected").data('price-mod'));
+      optionTotal += parseFloat($(this).find(":selected").data('price-mod'));
     }
   });
